Make write to Tx_HI start timer in one-shot mode
Previously, if Tx was stopped, and a write is done Tx_HI, then the
latch is copied to the counter.

With this change, if Tx is in one-shot mode, then a write to Tx_HI will
copy the latch to the counter, regardless if the timer is running
or not, and moreover the timer will start.